Rearrange code in config directory to match file structure in branch #691.
