+++ galsim/config/image.py
@@ -18,6 +18,9 @@
 
 import galsim
 
+from .image_scattered import *
+from .image_tiled import *
+
 # The items in each tuple are:
 #   - The function to call to build the image
 #   - The function to call to get the number of objects that will be built
@@ -328,6 +331,21 @@ def BuildImage(config, logger=None, image_num=0, obj_num=0,
     return all_images
 
 
+def GetNObjForImage(config, image_num):
+    if 'image' in config and 'type' in config['image']:
+        image_type = config['image']['type']
+    else:
+        image_type = 'Single'
+
+    # Check that the type is valid
+    if image_type not in valid_image_types:
+        raise AttributeError("Invalid image.type=%s."%type)
+
+    nobj_func = eval(valid_image_types[image_type][1])
+
+    return nobj_func(config,image_num)
+
+
 def _set_image_origin(config, convention):
     """Set `config['image_origin']` appropriately based on the provided `convention`.
     """
@@ -417,563 +435,7 @@ def BuildSingleImage(config, logger=None, image_num=0, obj_num=0,
                                                      # last two elements of the BuildSingleStamp
                                                      # return tuple
 
-
-def BuildTiledImage(config, logger=None, image_num=0, obj_num=0,
-                    make_psf_image=False, make_weight_image=False, make_badpix_image=False):
-    """
-    Build an Image consisting of a tiled array of postage stamps.
-
-    @param config              A configuration dict.
-    @param logger              If given, a logger object to log progress. [default: None]
-    @param image_num           If given, the current `image_num`. [default: 0]
-    @param obj_num             If given, the current `obj_num`. [default: 0]
-    @param make_psf_image      Whether to make `psf_image`. [default: False]
-    @param make_weight_image   Whether to make `weight_image`. [default: False]
-    @param make_badpix_image   Whether to make `badpix_image`. [default: False]
-
-    @returns the tuple `(image, psf_image, weight_image, badpix_image)`.
-
-    Note: All 4 Images are always returned in the return tuple,
-          but the latter 3 might be None depending on the parameters make_*_image.    
-    """
-    config['index_key'] = 'image_num'
-    config['image_num'] = image_num
-    config['obj_num'] = obj_num
-
-    if logger:
-        logger.debug('image %d: BuildTiledImage: image, obj = %d,%d',image_num,image_num,obj_num)
-
-    if 'random_seed' in config['image'] and not isinstance(config['image']['random_seed'],dict):
-        first = galsim.config.ParseValue(config['image'], 'random_seed', config, int)[0]
-        config['image']['random_seed'] = { 'type' : 'Sequence', 'first' : first }
-
-    ignore = [ 'random_seed', 'draw_method', 'noise', 'pixel_scale', 'wcs', 'nproc',
-               'sky_level', 'sky_level_pixel',
-               'retry_failures', 'image_pos', 'n_photons', 'wmult', 'offset', 'gsparams' ]
-    req = { 'nx_tiles' : int , 'ny_tiles' : int }
-    opt = { 'stamp_size' : int , 'stamp_xsize' : int , 'stamp_ysize' : int ,
-            'border' : int , 'xborder' : int , 'yborder' : int ,
-            'nproc' : int , 'index_convention' : str, 'order' : str }
-    params = galsim.config.GetAllParams(
-        config['image'], 'image', config, req=req, opt=opt, ignore=ignore)[0]
-
-    nx_tiles = params['nx_tiles']
-    ny_tiles = params['ny_tiles']
-    nobjects = nx_tiles * ny_tiles
-    config['nx_tiles'] = nx_tiles
-    config['ny_tiles'] = ny_tiles
-    if logger:
-        logger.debug('image %d: n_tiles = %d, %d',image_num,nx_tiles,ny_tiles)
-
-    stamp_size = params.get('stamp_size',0)
-    stamp_xsize = params.get('stamp_xsize',stamp_size)
-    stamp_ysize = params.get('stamp_ysize',stamp_size)
-    config['tile_xsize'] = stamp_xsize
-    config['tile_ysize'] = stamp_ysize
-
-    if (stamp_xsize == 0) or (stamp_ysize == 0):
-        raise AttributeError(
-            "Both image.stamp_xsize and image.stamp_ysize need to be defined and != 0.")
-
-    border = params.get("border",0)
-    xborder = params.get("xborder",border)
-    yborder = params.get("yborder",border)
-
-    do_noise = xborder >= 0 and yborder >= 0
-    # TODO: Note: if one of these is < 0 and the other is > 0, then
-    #       this will add noise to the border region.  Not exactly the 
-    #       design, but I didn't bother to do the bookkeeping right to 
-    #       make the borders pure 0 in that case.
-    
-    full_xsize = (stamp_xsize + xborder) * nx_tiles - xborder
-    full_ysize = (stamp_ysize + yborder) * ny_tiles - yborder
-
-    # If image_force_xsize and image_force_ysize were set in config, make sure it matches.
-    if ( ('image_force_xsize' in config and full_xsize != config['image_force_xsize']) or
-         ('image_force_ysize' in config and full_ysize != config['image_force_ysize']) ):
-        raise ValueError(
-            "Unable to reconcile required image xsize and ysize with provided "+
-            "nx_tiles=%d, ny_tiles=%d, "%(nx_tiles,ny_tiles) +
-            "xborder=%d, yborder=%d\n"%(xborder,yborder) +
-            "Calculated full_size = (%d,%d) "%(full_xsize,full_ysize)+
-            "!= required (%d,%d)."%(config['image_force_xsize'],config['image_force_ysize']))
-    config['image_xsize'] = full_xsize
-    config['image_ysize'] = full_ysize
-    if logger:
-        logger.debug('image %d: image_size = %d, %d',image_num,full_xsize,full_ysize)
-
-    convention = params.get('index_convention','1')
-    _set_image_origin(config,convention)
-    if logger:
-        logger.debug('image %d: image_origin = %s',image_num,str(config['image_origin']))
-        logger.debug('image %d: image_center = %s',image_num,str(config['image_center']))
-
-    wcs = galsim.config.BuildWCS(config, logger)
-
-    # Set the rng to use for image stuff.
-    if 'random_seed' in config['image']:
-        # Technically obj_num+nobjects will be the index of the random seed used for the next 
-        # image's first object (if there is a next image).  But I don't think that will have 
-        # any adverse effects.
-        config['obj_num'] = obj_num + nobjects
-        config['index_key'] = 'obj_num'
-        seed = galsim.config.ParseValue(config['image'], 'random_seed', config, int)[0]
-        config['index_key'] = 'image_num'
-        if logger:
-            logger.debug('image %d: seed = %d',image_num,seed)
-        rng = galsim.BaseDeviate(seed)
-    else:
-        rng = galsim.BaseDeviate()
-    config['rng'] = rng
-
-    # Make a list of ix,iy values according to the specified order:
-    order = params.get('order','row').lower()
-    if order.startswith('row'):
-        ix_list = [ ix for iy in range(ny_tiles) for ix in range(nx_tiles) ]
-        iy_list = [ iy for iy in range(ny_tiles) for ix in range(nx_tiles) ]
-    elif order.startswith('col'):
-        ix_list = [ ix for ix in range(nx_tiles) for iy in range(ny_tiles) ]
-        iy_list = [ iy for ix in range(nx_tiles) for iy in range(ny_tiles) ]
-    elif order.startswith('rand'):
-        ix_list = [ ix for ix in range(nx_tiles) for iy in range(ny_tiles) ]
-        iy_list = [ iy for ix in range(nx_tiles) for iy in range(ny_tiles) ]
-        galsim.random.permute(rng, ix_list, iy_list)
-        
-    # Define a 'image_pos' field so the stamps can set their position appropriately in case
-    # we need it for PowerSpectum or NFWHalo.
-    x0 = (stamp_xsize-1)/2. + config['image_origin'].x
-    y0 = (stamp_ysize-1)/2. + config['image_origin'].y
-    dx = stamp_xsize + xborder
-    dy = stamp_ysize + yborder
-    config['image']['image_pos'] = { 
-        'type' : 'XY' ,
-        'x' : { 'type' : 'List',
-                'items' : [ x0 + ix*dx for ix in ix_list ]
-              },
-        'y' : { 'type' : 'List',
-                'items' : [ y0 + iy*dy for iy in iy_list ]
-              }
-    }
-
-    nproc = params.get('nproc',1)
-
-    full_image = galsim.ImageF(full_xsize, full_ysize)
-    full_image.setOrigin(config['image_origin'])
-    full_image.wcs = wcs
-    full_image.setZero()
-
-    if make_psf_image:
-        full_psf_image = galsim.ImageF(full_image.bounds, wcs=wcs)
-        full_psf_image.setZero()
-    else:
-        full_psf_image = None
-
-    if make_weight_image:
-        full_weight_image = galsim.ImageF(full_image.bounds, wcs=wcs)
-        full_weight_image.setZero()
-    else:
-        full_weight_image = None
-
-    if make_badpix_image:
-        full_badpix_image = galsim.ImageS(full_image.bounds, wcs=wcs)
-        full_badpix_image.setZero()
-    else:
-        full_badpix_image = None
-
-    # Sometimes an input field needs to do something special at the start of an image.
-    if 'input' in config:
-        for key in [ k for k in galsim.config.valid_input_types.keys() if k in config['input'] ]:
-            if galsim.config.valid_input_types[key][4]:
-                assert key in config
-                fields = config['input'][key]
-                if not isinstance(fields, list):
-                    fields = [ fields ]
-                input_objs = config[key]
-
-                for i in range(len(fields)):
-                    field = fields[i]
-                    input_obj = input_objs[i]
-                    func = eval(galsim.config.valid_input_types[key][4])
-                    func(input_obj, field, config)
-
-    stamp_images = galsim.config.BuildStamps(
-            nobjects=nobjects, config=config,
-            nproc=nproc, logger=logger, obj_num=obj_num,
-            xsize=stamp_xsize, ysize=stamp_ysize, do_noise=do_noise,
-            make_psf_image=make_psf_image,
-            make_weight_image=make_weight_image,
-            make_badpix_image=make_badpix_image)
-
-    images = stamp_images[0]
-    psf_images = stamp_images[1]
-    weight_images = stamp_images[2]
-    badpix_images = stamp_images[3]
-    current_vars = stamp_images[4]
-
-    max_current_var = 0
-    for k in range(nobjects):
-        # This is our signal that the object was skipped.
-        if not images[k].bounds.isDefined(): continue
-        if False:
-            logger.debug('image %d: full bounds = %s',image_num,str(full_image.bounds))
-            logger.debug('image %d: stamp %d bounds = %s',image_num,k,str(images[k].bounds))
-        assert full_image.bounds.includes(images[k].bounds)
-        b = images[k].bounds
-        full_image[b] += images[k]
-        if make_psf_image:
-            full_psf_image[b] += psf_images[k]
-        if make_weight_image:
-            full_weight_image[b] += weight_images[k]
-        if make_badpix_image:
-            full_badpix_image[b] |= badpix_images[k]
-        if current_vars[k] > max_current_var: max_current_var = current_vars[k]
-
-    # Mark that we are no longer doing a single galaxy by deleting image_pos from config top 
-    # level, so it cannot be used for things like wcs.pixelArea(image_pos).  
-    if 'image_pos' in config: del config['image_pos']
-
-    # If didn't do noise above in the stamps, then need to do it here.
-    if not do_noise:
-        if 'noise' in config['image']:
-            # If we didn't apply noise in each stamp, then we need to apply it now.
-            draw_method = galsim.config.GetCurrentValue(config['image'],'draw_method')
-
-            if max_current_var > 0:
-                import numpy
-                # Then there was whitening applied in the individual stamps.
-                # But there could be a different variance in each postage stamp, so the first
-                # thing we need to do is bring everything up to a common level.
-                noise_image = galsim.ImageF(full_image.bounds)
-                for k in range(nobjects): noise_image[images[k].bounds] += current_vars[k]
-                # Update this, since overlapping postage stamps may have led to a larger 
-                # value in some pixels.
-                max_current_var = numpy.max(noise_image.array)
-                # Figure out how much noise we need to add to each pixel.
-                noise_image = max_current_var - noise_image
-                # Add it.
-                full_image.addNoise(galsim.VariableGaussianNoise(rng,noise_image))
-            # Now max_current_var is how much noise is in each pixel.
-
-            config['rng'] = rng
-            galsim.config.AddNoise(
-                config,draw_method,full_image,full_weight_image,max_current_var,logger)
-
-        else:
-            # If we aren't doing noise, we still may need to add a non-zero sky_level.
-            # The same noise function does this with the 'skip' draw method.
-            galsim.config.AddNoise(
-                config,'skip',full_image,full_weight_image,max_current_var,logger)
-
-    return full_image, full_psf_image, full_weight_image, full_badpix_image
-
-
-def BuildScatteredImage(config, logger=None, image_num=0, obj_num=0,
-                        make_psf_image=False, make_weight_image=False, make_badpix_image=False):
-    """
-    Build an Image containing multiple objects placed at arbitrary locations.
-
-    @param config              A configuration dict.
-    @param logger              If given, a logger object to log progress. [default None]
-    @param image_num           If given, the current `image_num` [default: 0]
-    @param obj_num             If given, the current `obj_num` [default: 0]
-    @param make_psf_image      Whether to make `psf_image`. [default: False]
-    @param make_weight_image   Whether to make `weight_image`. [default: False]
-    @param make_badpix_image   Whether to make `badpix_image`. [default: False]
-
-    @returns the tuple `(image, psf_image, weight_image, badpix_image)`.
-
-    Note: All 4 Images are always returned in the return tuple,
-          but the latter 3 might be None depending on the parameters make_*_image.    
-    """
-    config['index_key'] = 'image_num'
-    config['image_num'] = image_num
-    config['obj_num'] = obj_num
-
-    if logger:
-        logger.debug('image %d: BuildScatteredImage: image, obj = %d,%d',
-                     image_num,image_num,obj_num)
-
-    if 'random_seed' in config['image'] and not isinstance(config['image']['random_seed'],dict):
-        first = galsim.config.ParseValue(config['image'], 'random_seed', config, int)[0]
-        config['image']['random_seed'] = { 'type' : 'Sequence', 'first' : first }
-
-    nobjects = GetNObjForScatteredImage(config,image_num)
-    if logger:
-        logger.debug('image %d: nobj = %d',image_num,nobjects)
-
-    ignore = [ 'random_seed', 'draw_method', 'noise', 'pixel_scale', 'wcs', 'nproc',
-               'sky_level', 'sky_level_pixel',
-               'retry_failures', 'image_pos', 'world_pos', 'n_photons', 'wmult', 'offset', 
-               'stamp_size', 'stamp_xsize', 'stamp_ysize', 'gsparams', 'nobjects' ]
-    opt = { 'size' : int , 'xsize' : int , 'ysize' : int , 
-            'nproc' : int , 'index_convention' : str }
-    params = galsim.config.GetAllParams(
-        config['image'], 'image', config, opt=opt, ignore=ignore)[0]
-
-    # Special check for the size.  Either size or both xsize and ysize is required.
-    if 'size' not in params:
-        if 'xsize' not in params or 'ysize' not in params:
-            raise AttributeError(
-                "Either attribute size or both xsize and ysize required for image.type=Scattered")
-        full_xsize = params['xsize']
-        full_ysize = params['ysize']
-    else:
-        if 'xsize' in params:
-            raise AttributeError(
-                "Attributes xsize is invalid if size is set for image.type=Scattered")
-        if 'ysize' in params:
-            raise AttributeError(
-                "Attributes ysize is invalid if size is set for image.type=Scattered")
-        full_xsize = params['size']
-        full_ysize = params['size']
-
-
-    # If image_force_xsize and image_force_ysize were set in config, make sure it matches.
-    if ( ('image_force_xsize' in config and full_xsize != config['image_force_xsize']) or
-         ('image_force_ysize' in config and full_ysize != config['image_force_ysize']) ):
-        raise ValueError(
-            "Unable to reconcile required image xsize and ysize with provided "+
-            "xsize=%d, ysize=%d, "%(full_xsize,full_ysize))
-    config['image_xsize'] = full_xsize
-    config['image_ysize'] = full_ysize
-
-    convention = params.get('index_convention','1')
-    _set_image_origin(config,convention)
-    if logger:
-        logger.debug('image %d: image_origin = %s',image_num,str(config['image_origin']))
-        logger.debug('image %d: image_center = %s',image_num,str(config['image_center']))
-
-    wcs = galsim.config.BuildWCS(config, logger)
-
-    # Set the rng to use for image stuff.
-    if 'random_seed' in config['image']:
-        # Technically obj_num+nobjects will be the index of the random seed used for the next 
-        # image's first object (if there is a next image).  But I don't think that will have 
-        # any adverse effects.
-        config['obj_num'] = obj_num + nobjects
-        config['index_key'] = 'obj_num'
-        seed = galsim.config.ParseValue(config['image'], 'random_seed', config, int)[0]
-        config['index_key'] = 'image_num'
-        if logger:
-            logger.debug('image %d: seed = %d',image_num,seed)
-        rng = galsim.BaseDeviate(seed)
-    else:
-        rng = galsim.BaseDeviate()
-    config['rng'] = rng
-
-    if 'image_pos' in config['image'] and 'world_pos' in config['image']:
-        raise AttributeError("Both image_pos and world_pos specified for Scattered image.")
-
-    if 'image_pos' not in config['image'] and 'world_pos' not in config['image']:
-        xmin = config['image_origin'].x
-        xmax = xmin + full_xsize-1
-        ymin = config['image_origin'].y
-        ymax = ymin + full_ysize-1
-        config['image']['image_pos'] = { 
-            'type' : 'XY' ,
-            'x' : { 'type' : 'Random' , 'min' : xmin , 'max' : xmax },
-            'y' : { 'type' : 'Random' , 'min' : ymin , 'max' : ymax }
-        }
-
-    nproc = params.get('nproc',1)
-
-    full_image = galsim.ImageF(full_xsize, full_ysize)
-    full_image.setOrigin(config['image_origin'])
-    full_image.wcs = wcs
-    full_image.setZero()
-
-    if make_psf_image:
-        full_psf_image = galsim.ImageF(full_image.bounds, wcs=wcs)
-        full_psf_image.setZero()
-    else:
-        full_psf_image = None
-
-    if make_weight_image:
-        full_weight_image = galsim.ImageF(full_image.bounds, wcs=wcs)
-        full_weight_image.setZero()
-    else:
-        full_weight_image = None
-
-    if make_badpix_image:
-        full_badpix_image = galsim.ImageS(full_image.bounds, wcs=wcs)
-        full_badpix_image.setZero()
-    else:
-        full_badpix_image = None
-
-    # Sometimes an input field needs to do something special at the start of an image.
-    if 'input' in config:
-        for key in [ k for k in galsim.config.valid_input_types.keys() if k in config['input'] ]:
-            if galsim.config.valid_input_types[key][4]:
-                assert key in config
-                fields = config['input'][key]
-                if not isinstance(fields, list):
-                    fields = [ fields ]
-                input_objs = config[key]
-
-                for i in range(len(fields)):
-                    field = fields[i]
-                    input_obj = input_objs[i]
-                    func = eval(galsim.config.valid_input_types[key][4])
-                    func(input_obj, field, config)
-
-    stamp_images = galsim.config.BuildStamps(
-            nobjects=nobjects, config=config,
-            nproc=nproc, logger=logger,obj_num=obj_num, do_noise=False,
-            make_psf_image=make_psf_image,
-            make_weight_image=make_weight_image,
-            make_badpix_image=make_badpix_image)
-
-    images = stamp_images[0]
-    psf_images = stamp_images[1]
-    weight_images = stamp_images[2]
-    badpix_images = stamp_images[3]
-    current_vars = stamp_images[4]
-
-    max_current_var = 0.
-    for k in range(nobjects):
-        # This is our signal that the object was skipped.
-        if not images[k].bounds.isDefined(): continue
-        bounds = images[k].bounds & full_image.bounds
-        if False:
-            logger.debug('image %d: full bounds = %s',image_num,str(full_image.bounds))
-            logger.debug('image %d: stamp %d bounds = %s',image_num,k,str(images[k].bounds))
-            logger.debug('image %d: Overlap = %s',image_num,str(bounds))
-        if bounds.isDefined():
-            full_image[bounds] += images[k][bounds]
-            if make_psf_image:
-                full_psf_image[bounds] += psf_images[k][bounds]
-            if make_weight_image:
-                full_weight_image[bounds] += weight_images[k][bounds]
-            if make_badpix_image:
-                full_badpix_image[bounds] |= badpix_images[k][bounds]
-        else:
-            if logger:
-                logger.warn(
-                    "Object centered at (%d,%d) is entirely off the main image,\n"%(
-                        images[k].bounds.center().x, images[k].bounds.center().y) +
-                    "whose bounds are (%d,%d,%d,%d)."%(
-                        full_image.bounds.xmin, full_image.bounds.xmax,
-                        full_image.bounds.ymin, full_image.bounds.ymax))
-        if current_vars[k] > max_current_var: max_current_var = current_vars[k]
-
-    # Mark that we are no longer doing a single galaxy by deleting image_pos from config top 
-    # level, so it cannot be used for things like wcs.pixelArea(image_pos).  
-    if 'image_pos' in config: del config['image_pos']
-
-    if 'noise' in config['image']:
-        # Apply the noise to the full image
-        draw_method = galsim.config.GetCurrentValue(config['image'],'draw_method')
-        if max_current_var > 0:
-            import numpy
-            # Then there was whitening applied in the individual stamps.
-            # But there could be a different variance in each postage stamp, so the first
-            # thing we need to do is bring everything up to a common level.
-            noise_image = galsim.ImageF(full_image.bounds)
-            for k in range(nobjects): 
-                b = images[k].bounds & full_image.bounds
-                if b.isDefined(): noise_image[b] += current_vars[k]
-            # Update this, since overlapping postage stamps may have led to a larger 
-            # value in some pixels.
-            max_current_var = numpy.max(noise_image.array)
-            # Figure out how much noise we need to add to each pixel.
-            noise_image = max_current_var - noise_image
-            # Add it.
-            full_image.addNoise(galsim.VariableGaussianNoise(rng,noise_image))
-        # Now max_current_var is how much noise is in each pixel.
-
-        config['rng'] = rng
-        galsim.config.AddNoise(
-            config,draw_method,full_image,full_weight_image,max_current_var,logger)
-
-    else:
-        # If we aren't doing noise, we still may need to add a non-zero sky_level.
-        # The same noise function does this with the 'skip' draw method.
-        galsim.config.AddNoise(
-            config,'skip',full_image,full_weight_image,max_current_var,logger)
-
-    return full_image, full_psf_image, full_weight_image, full_badpix_image
-
-
-def GetNObjForImage(config, image_num):
-    if 'image' in config and 'type' in config['image']:
-        image_type = config['image']['type']
-    else:
-        image_type = 'Single'
-
-    # Check that the type is valid
-    if image_type not in valid_image_types:
-        raise AttributeError("Invalid image.type=%s."%type)
-
-    nobj_func = eval(valid_image_types[image_type][1])
-
-    return nobj_func(config,image_num)
-
 def GetNObjForSingleImage(config, image_num):
     return 1
-
-def GetNObjForScatteredImage(config, image_num):
-
-    config['index_key'] = 'image_num'
-    config['image_num'] = image_num
-
-    # Allow nobjects to be automatic based on input catalog
-    if 'nobjects' not in config['image']:
-        nobj = galsim.config.ProcessInputNObjects(config)
-        if nobj is None:
-            raise AttributeError("Attribute nobjects is required for image.type = Scattered")
-        return nobj
-    else:
-        nobj = galsim.config.ParseValue(config['image'],'nobjects',config,int)[0]
-        return nobj
-
-def GetNObjForTiledImage(config, image_num):
-    
-    config['index_key'] = 'image_num'
-    config['image_num'] = image_num
-
-    if 'nx_tiles' not in config['image'] or 'ny_tiles' not in config['image']:
-        raise AttributeError(
-            "Attributes nx_tiles and ny_tiles are required for image.type = Tiled")
-    nx = galsim.config.ParseValue(config['image'],'nx_tiles',config,int)[0]
-    ny = galsim.config.ParseValue(config['image'],'ny_tiles',config,int)[0]
-    return nx*ny
-
-def PowerSpectrumInit(ps, config, base):
-    if 'grid_spacing' in config:
-        grid_spacing = galsim.config.ParseValue(config, 'grid_spacing', base, float)[0]
-    elif 'tile_xsize' in base:
-        # Then we have a tiled image.  Can use the tile spacing as the grid spacing.
-        stamp_size = min(base['tile_xsize'], base['tile_ysize'])
-        # Note: we use the (max) pixel scale at the image center.  This isn't 
-        # necessarily optimal, but it seems like the best choice.
-        scale = base['wcs'].maxLinearScale(base['image_center'])
-        grid_spacing = stamp_size * scale
-    else:
-        raise AttributeError("power_spectrum.grid_spacing required for non-tiled images")
-
-    if 'tile_xsize' in base and base['tile_xsize'] == base['tile_ysize']:
-        # PowerSpectrum can only do a square FFT, so make it the larger of the two n's.
-        ngrid = max(base['nx_tiles'], base['ny_tiles'])
-        # Normally that's good, but if tiles aren't square, need to drop through to the
-        # second option.
-    else:
-        import math
-        image_size = max(base['image_xsize'], base['image_ysize'])
-        scale = base['wcs'].maxLinearScale(base['image_center'])
-        ngrid = int(math.ceil(image_size * scale / grid_spacing))
-
-    if 'interpolant' in config:
-        interpolant = galsim.config.ParseValue(config, 'interpolant', base, str)[0]
-    else:
-        interpolant = None
-
-    # We don't care about the output here.  This just builds the grid, which we'll
-    # access for each object using its position.
-    if base['wcs'].isCelestial():
-        world_center = galsim.PositionD(0,0)
-    else:
-        world_center = base['wcs'].toWorld(base['image_center'])
-    ps.buildGrid(grid_spacing=grid_spacing, ngrid=ngrid, center=world_center,
-                 rng=base['rng'], interpolant=interpolant)
