+++ galsim/config/stamp.py
@@ -504,7 +504,7 @@ def BuildSingleStamp(config, xsize=0, ysize=0,
                     logger.info('This is try %d/%d, so trying again.',itry+1,ntries)
                 # Need to remove the "current_val"s from the config dict.  Otherwise,
                 # the value generators will do a quick return with the cached value.
-                galsim.config.process.RemoveCurrent(config, keep_safe=True)
+                galsim.config.RemoveCurrent(config, keep_safe=True)
                 continue
 
     return im, psf_im, weight_im, badpix_im, current_var, t6-t1
